+++ webapp/src/pages/Dashboard.tsx
@@ -3,6 +3,7 @@ import noData from "assets/void.svg";
 import PerturbationTestingPreview from "components/Analysis/PerturbationTestingPreview";
 import PreviewCard from "components/Analysis/PreviewCard";
 import WarningsPreview from "components/Analysis/WarningsPreview";
+import ClassAnalysisTable from "components/ClassAnalysisTable";
 import Description from "components/Description";
 import Telescope from "components/Icons/Telescope";
 import Loading from "components/Loading";
@@ -14,6 +15,7 @@ import React from "react";
 import { Link, useParams } from "react-router-dom";
 import { getDatasetInfoEndpoint } from "services/api";
 import { isPipelineSelected } from "utils/helpers";
+import { classAnalysisDescription } from "./ClassAnalysis";
 import { performanceAnalysisDescription } from "./PerformanceAnalysis";
 import { behavioralTestingDescription } from "./PerturbationTestingSummary";
 import { smartTagsDescription } from "./SmartTags";
@@ -83,6 +85,15 @@ const Dashboard = () => {
           <Box height={DEFAULT_PREVIEW_CONTENT_HEIGHT}>
             <WarningsPreview jobId={jobId} />
           </Box>
+        </PreviewCard>
+      )}
+      {datasetInfo?.similarityAvailable && (
+        <PreviewCard
+          title="Class Analysis"
+          to={`/${jobId}/class_analysis${searchString}`}
+          description={classAnalysisDescription}
+        >
+          <ClassAnalysisTable jobId={jobId} pipeline={pipeline} />
         </PreviewCard>
       )}
       {isPipelineSelected(pipeline) && (